+++ project_templates/stack_package/{{cookiecutter.package_name}}/doc/conf.py
@@ -1,20 +1,14 @@
 """Sphinx configuration file for an LSST stack package.
 
 This configuration only affects single-package Sphinx documentation builds.
+For more information, see:
+https://developer.lsst.io/stack/building-single-package-docs.html
 """
 
-from documenteer.sphinxconfig.stackconf import build_package_configs
-{%- if cookiecutter.uses_python == 'True' %}
-import {{ cookiecutter.python_module }}
+from documenteer.conf.pipelinespkg import *
 
 
-_g = globals()
-_g.update(build_package_configs(
-    project_name='{{ cookiecutter.package_name }}',
-    version={{ cookiecutter.python_module }}.version.__version__))
-{% else %}
-
-
-_g = globals()
-_g.update(build_package_configs(project_name='{{ cookiecutter.package_name }}'))
-{% endif -%}
+project = "{{ cookiecutter.package_name }}"
+html_theme_options["logotext"] = project
+html_title = project
+html_short_title = project